Retry CMA allocations.
It looks like Linux pages migration code was never designed to be
deterministic or synchronous, there are multiple race conditions
between different parts of the code that make CMA allocation in one
step very likely to fail, especially for large memory ranges that
we need for Ducati.

Therefore, changing the allocation code to perform multiple allocation
attempts. To further increase the chances of the allocation to succeed
and to make things faster, the results of the previous allocation
attempts are kept - that is, all pages that are already isolated
stay isolated, so that retries are only for those pages that failed
isolation or migration in previous steps.

Additionally, there's a small delay between the steps so that the
chances of the other code to free the pages we need are higher.